Flip pi
Whoops.  For all this time, the pi footprint was backwards, as if this board was a hat mounting on top of the pi.  Flip it so the pi mounts face down on this board.  This will put the network and usb connections towards the rear panel, so move it forward as far as possible to make room.